requirements.txt
staticmap==0.5.4
tqdm
mega.py
requests==2.24.0
opencv_contrib_python
scikit_image==0.17.2
scipy==1.1.0
matplotlib
numpy==1.19.4
utm
Pillow==8.2.0
protobuf==3.15.8
googledrivedownloader
